+++ src/main/java/com/nextech/server/v1/domain/members/service/impl/CurrentWardMembersInquiryServiceImpl.java
@@ -2,7 +2,7 @@ package com.nextech.server.v1.domain.members.service.impl;
 
 import com.nextech.server.v1.domain.members.dto.response.MembersInquiryResponse;
 import com.nextech.server.v1.domain.members.entity.Members;
-import com.nextech.server.v1.domain.members.repository.MemberRepository;
+import com.nextech.server.v1.global.members.repository.MemberRepository;
 import com.nextech.server.v1.domain.members.service.CurrentWardMembersInquiryService;
 import com.nextech.server.v1.global.exception.RelationNullException;
 import com.nextech.server.v1.global.members.dto.response.MembersInquiryListResponse;
