+++ src/main/java/com/nextech/server/v1/domain/members/service/impl/RoleChangeServiceImpl.java
@@ -2,7 +2,7 @@ package com.nextech.server.v1.domain.members.service.impl;
 
 import com.nextech.server.v1.domain.auth.service.PasswordValidationService;
 import com.nextech.server.v1.domain.members.entity.Members;
-import com.nextech.server.v1.domain.members.repository.MemberRepository;
+import com.nextech.server.v1.global.members.repository.MemberRepository;
 import com.nextech.server.v1.domain.members.service.RoleChangeService;
 import com.nextech.server.v1.global.enums.Roles;
 import com.nextech.server.v1.global.exception.DuplicateRoleAssignmentException;
